# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API GEST_RV_CHUN",
        "description": "Documentation des endpoints API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Serveur courant"
        }
    ],
    "paths": {
...
        "/dossiers/{id}": {
            "get": {
                "tags": [
                    "Dossiers"
                ],
                "summary": "Consulter un dossier médical",
                "operationId": "8a1ee896b295d3602dcfc17239897b79",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DossierMedicalResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Dossiers"
                ],
                "summary": "Mettre à jour un dossier médical",
                "operationId": "ac7d88cd1c0594fbf6275a3f13a7b8cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DossierMedicalUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DossierMedicalResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DossierMedical": {
                "properties": {
                    "identification": {
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "numero_dossier": {
                                "type": "string"
                            },
                            "created_at": {
                                "type": "string"
                            },
                            "updated_at": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "patient": {
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "matricule": {
                                "type": "string"
                            },
                            "nom": {
                                "type": "string"
                            },
                            "prenom": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            },
                            "telephone": {
                                "type": "string"
                            },
                            "date_naissance": {
                                "type": "string"
                            },
                            "age": {
                                "type": "integer"
                            },
                            "adresse": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "informations_medicales": {
                        "properties": {
                            "groupe_sanguin": {
                                "type": "string",
                                "nullable": true
                            },
                            "antecedents_medicaux": {
                                "type": "string",
                                "nullable": true
                            },
                            "antecedents_chirurgicaux": {
                                "type": "string",
                                "nullable": true
                            },
                            "antecedents_familiaux": {
                                "type": "string",
                                "nullable": true
                            },
                            "allergies": {
                                "type": "string",
                                "nullable": true
                            },
                            "maladies_chroniques": {
                                "type": "string",
                                "nullable": true
                            },
                            "traitements_en_cours": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "statistiques": {
                        "properties": {
                            "nombre_consultations": {
                                "type": "integer"
                            },
                            "nombre_rendez_vous_total": {
                                "type": "integer"
                            },
                            "nombre_rendez_vous_annules": {
                                "type": "integer"
                            },
                            "date_derniere_consultation": {
                                "type": "string",
                                "nullable": true
                            },
                            "date_premier_rendez_vous": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "derniere_consultation": {
                        "properties": {
                            "date_heure": {
                                "type": "string"
                            },
                            "statut": {
                                "type": "string"
                            },
                            "medecin": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "service": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "sumptomes": {
                                "type": "string"
                            },
                            "tension_artielle": {
                                "type": "string",
                                "nullable": true
                            },
                            "poids": {
                                "type": "number",
                                "nullable": true
                            },
                            "temperature": {
                                "type": "number",
                                "nullable": true
                            },
                            "diagnostic": {
                                "type": "string",
                                "nullable": true
                            },
                            "traitement": {
                                "type": "string",
                                "nullable": true
                            },
                            "observations": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Non authentifie"
                    },
                    "data": {
                        "nullable": true
                    },
                    "errors": {
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "authentication"
                            },
                            "detail": {
                                "type": "string",
                                "example": "Authentification requise pour acceder a cette ressource."
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "DossierMedicalResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Dossier médical"
                    },
                    "data": {
                        "$ref": "#/components/schemas/DossierMedical"
                    },
                    "errors": {
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "DossierMedicalUpdateRequest": {
                "properties": {
                    "groupe_sanguin": {
                        "type": "string",
                        "example": "O+"
                    },
                    "antecedents_medicaux": {
                        "type": "string",
                        "example": "Hypertension"
                    },
                    "antecedents_chirurgicaux": {
                        "type": "string",
                        "example": "Aucun"
                    },
                    "antecedents_familiaux": {
                        "type": "string",
                        "example": "Diabète"
                    },
                    "allergies": {
                        "type": "string",
                        "example": "Aucune"
                    },
                    "maladies_chroniques": {
                        "type": "string",
                        "example": "Asthme"
                    },
                    "traitements_en_cours": {
                        "type": "string",
                        "example": "Traitement A"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}